requirements.txt
# requirements.txt

fastmcp
python-dotenv
psycopg2-binary